added permissions hub module
