+++ dex/farm-with-locked-rewards/Cargo.toml
@@ -41,6 +41,9 @@ path = "../../common/modules/utils"
 [dependencies.permissions_module]
 path = "../../common/modules/permissions_module"
 
+[dependencies.permissions_hub_module]
+path = "../../common/modules/permissions_hub_module"
+
 [dependencies.sc_whitelist_module]
 path = "../../common/modules/sc_whitelist_module"
