+++ dex/farm/Cargo.toml
@@ -38,6 +38,9 @@ path = "../../common/modules/pausable"
 [dependencies.permissions_module]
 path = "../../common/modules/permissions_module"
 
+[dependencies.permissions_hub_module]
+path = "../../common/modules/permissions_hub_module"
+
 [dependencies.sc_whitelist_module]
 path = "../../common/modules/sc_whitelist_module"
