+++ farm-staking/farm-staking-proxy/Cargo.toml
@@ -66,6 +66,9 @@ path = "../../energy-integration/common-modules/energy-query"
 [dependencies.permissions-hub]
 path = "../../dex/permissions-hub"
 
+[dependencies.permissions_hub_module]
+path = "../../common/modules/permissions_hub_module"
+
 [dev-dependencies]
 num-bigint = "0.4.2"
