+++ farm-staking/farm-staking/Cargo.toml
@@ -53,6 +53,9 @@ path = "../../common/modules/pausable"
 [dependencies.permissions_module]
 path = "../../common/modules/permissions_module"
 
+[dependencies.permissions_hub_module]
+path = "../../common/modules/permissions_hub_module"
+
 [dependencies.sc_whitelist_module]
 path = "../../common/modules/sc_whitelist_module"
