+++ farm-staking/farm-staking/tests/farm_staking_setup/mod.rs
@@ -27,6 +27,7 @@ use farm_staking::*;
 use farm_token::FarmTokenModule;
 use pausable::{PausableModule, State};
 use permissions_hub::PermissionsHub;
+use permissions_hub_module::PermissionsHubModule;
 use rewards::RewardsModule;
 
 pub static REWARD_TOKEN_ID: &[u8] = b"RIDE-abcdef"; // reward token ID
